+++ src/assets/scripts/airport/StandardRoute/RouteSegmentCollection.js
@@ -31,8 +31,9 @@ export default class RouteSegmentCollection extends BaseCollection {
     /**
      * @constructor
      * @param routeSegments {object}
+     * @param fixCollection {FixCollection}
      */
-    constructor(routeSegments) {
+    constructor(routeSegments, fixCollection) {
         super(routeSegments);
 
         if (typeof routeSegments === 'undefined' || !_isObject(routeSegments) || _isArray(routeSegments)) {
@@ -48,7 +49,7 @@ export default class RouteSegmentCollection extends BaseCollection {
          */
         this.name = '';
 
-        return this._init(routeSegments);
+        return this._init(routeSegments, fixCollection);
     }
 
     /**
@@ -67,12 +68,13 @@ export default class RouteSegmentCollection extends BaseCollection {
      * @for RouteSegmentCollection
      * @method _init
      * @param routeSegments {object}
+     * @param fixCollection {FixCollection}
      * @chainable
      * @private
      */
-    _init(routeSegments) {
+    _init(routeSegments, fixCollection) {
         _forEach(routeSegments, (routeWaypoints, key) => {
-            const routeSegmentModel = new RouteSegmentModel(key, routeWaypoints);
+            const routeSegmentModel = new RouteSegmentModel(key, routeWaypoints, fixCollection);
 
             this._addSegmentToCollection(routeSegmentModel);
         });