+++ src/assets/scripts/airport/StandardRoute/StandardRouteModel.js
@@ -52,8 +52,9 @@ export default class StandardRouteModel extends BaseModel {
      *
      * @constructor
      * @param standardRoute {object}
+     * @param fixCollection {FixCollection}
      */
-    constructor(standardRoute) {
+    constructor(standardRoute, fixCollection) {
         super();
 
         if (!_isObject(standardRoute) || _isArray(standardRoute)) {
@@ -163,7 +164,7 @@ export default class StandardRouteModel extends BaseModel {
          */
         this._entryCollection = null;
 
-        return this._init(standardRoute);
+        return this._init(standardRoute, fixCollection);
     }
 
     /**
@@ -172,9 +173,10 @@ export default class StandardRouteModel extends BaseModel {
      * @for StandardRouteModel
      * @method _init
      * @param standardRoute {object}
+     * @param fixCollection {FixCollection}
      * @private
      */
-    _init(standardRoute) {
+    _init(standardRoute, fixCollection) {
         this.icao = standardRoute.icao;
         this.name = standardRoute.name;
         this.draw = standardRoute.draw;
@@ -182,10 +184,10 @@ export default class StandardRouteModel extends BaseModel {
         this.body = standardRoute.body;
         this.exitPoints = _get(standardRoute, 'exitPoints', {});
         this.entryPoints = _get(standardRoute, 'entryPoints', {});
-        this._runwayCollection = this._buildSegmentCollection(standardRoute.rwy);
-        this._bodySegmentModel = this._buildSegmentModel(standardRoute.body);
-        this._exitCollection = this._buildSegmentCollection(standardRoute.exitPoints);
-        this._entryCollection = this._buildSegmentCollection(standardRoute.entryPoints);
+        this._runwayCollection = this._buildSegmentCollection(standardRoute.rwy, fixCollection);
+        this._bodySegmentModel = this._buildSegmentModel(standardRoute.body, fixCollection);
+        this._exitCollection = this._buildSegmentCollection(standardRoute.exitPoints, fixCollection);
+        this._entryCollection = this._buildSegmentCollection(standardRoute.entryPoints, fixCollection);
     }
 
     /**
@@ -309,11 +311,12 @@ export default class StandardRouteModel extends BaseModel {
      * @for StandardRouteModel
      * @method _buildSegmentModel
      * @param segmentFixList {array}
+     * @param fixCollection {FixCollection}
      * @return segmentModel {SegmentModel}
      * @private
      */
-    _buildSegmentModel(segmentFixList) {
-        const segmentModel = new RouteSegmentModel('body', segmentFixList);
+    _buildSegmentModel(segmentFixList, fixCollection) {
+        const segmentModel = new RouteSegmentModel('body', segmentFixList, fixCollection);
 
         return segmentModel;
     }
@@ -324,17 +327,18 @@ export default class StandardRouteModel extends BaseModel {
      * @for StandardRouteModel
      * @method _buildSegmentCollection
      * @param segment {object}
+     * @param fixCollection {FixCollection}
      * @return segmentCollection {SegmentCollection}
      * @private
      */
-    _buildSegmentCollection(segment) {
+    _buildSegmentCollection(segment, fixCollection) {
         // SIDS have `exitPoints` while STARs have `entryPoints`. one or the other will be `undefined`
         // depending on the route type.
         if (typeof segment === 'undefined' || _isEmpty(segment)) {
             return null;
         }
 
-        const segmentCollection = new RouteSegmentCollection(segment);
+        const segmentCollection = new RouteSegmentCollection(segment, fixCollection);
 
         return segmentCollection;
     }