+++ src/assets/scripts/airport/StandardRoute/StandardRouteWaypointModel.js
@@ -32,13 +32,16 @@ export default class StandardRouteWaypointModel extends BaseModel {
      *
      * @constructor
      * @param routeWaypoint {array|string}
+     * @param fixCollection {FixCollection}
      */
-    constructor(routeWaypoint) {
+    constructor(routeWaypoint, fixCollection) {
         super(routeWaypoint);
 
         if (typeof routeWaypoint === 'undefined') {
             return this;
         }
+
+        this._fixCollection = fixCollection;
 
         /**
          * Name of the fix
@@ -146,7 +149,7 @@ export default class StandardRouteWaypointModel extends BaseModel {
          */
         this.previousStandardWaypointName = '';
 
-        return this._init(routeWaypoint)
+        return this._init(routeWaypoint, fixCollection)
                    .clonePoisitonFromFix();
     }
 
@@ -191,10 +194,11 @@ export default class StandardRouteWaypointModel extends BaseModel {
      * @for StandardRouteWaypointModel
      * @method _init
      * @param routeWaypoint {array|string}
+     * @param fixCollection {FixCollection}
      * @chainable
      * @private
      */
-    _init(routeWaypoint) {
+    _init(routeWaypoint, fixCollection) {
         // if we receive a string, this fix doesnt have any restrictions so we only need to set `name`
         if (typeof routeWaypoint === 'string') {
             this.name = routeWaypoint;
@@ -218,6 +222,7 @@ export default class StandardRouteWaypointModel extends BaseModel {
      * @method reset
      */
     reset() {
+        this._fixCollection = null;
         this.name = '';
         this._restrictions = null;
         this._alititude = -1000;
@@ -236,7 +241,7 @@ export default class StandardRouteWaypointModel extends BaseModel {
      * @private
      */
     clonePoisitonFromFix() {
-        const fixModel = FixCollection.findFixByName(this.name);
+        const fixModel = this._fixCollection.findFixByName(this.name);
 
         if (!fixModel) {
             console.warn(`The following fix was not found in the list of fixes for this Airport: ${this.name}`);